requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
pandas>=1.3.0
gdown>=4.5.0
opencv-python>=4.5.0
tensorflow-addons>=0.17.0
